baas(S1.8): address cpp + security review findings
SDK (game-facing, incl. wasm32 where long is 32-bit):
- JSON parser: recursion depth cap (stack-overflow DoS on malicious/MITM body),
  reject invalid/lone UTF-16 surrogates (was emitting garbage UTF-8).
- 64-bit wire integers: json Value + Session/Rank/Entry + submit() use long long,
  strtoll (was long/strtol -> silent truncation of user_id/score on wasm32/Win).
- Transport: cap response size (curl + emscripten) and null-check curl handles.
Backend:
- password.cc: sodium_memzero the hash buffer after use.
- lb_controller: try/catch parity in top()/me() (match submit/auth handlers).
- colony_scene: my_score_ long long + %lld formats.
Tests: sdk_client gains parser-hardening cases (deep nesting, bad surrogates,
2^53+1 no-truncation). All 19 suites green; ASan/UBSan clean; web SDK builds.

+++ sdk/cpp/src/client.cc
@@ -87,7 +87,7 @@ void Client::Auth::registerUser(const std::string& email, const std::string& pas
 }
 
 // -------- Leaderboard --------
-void Client::Leaderboard::submit(long value, std::function<void(Result<Rank>)> cb) {
+void Client::Leaderboard::submit(long long value, std::function<void(Result<Rank>)> cb) {
     const std::string body = "{\"value\":" + std::to_string(value) + "}";
     c_->request<Rank>("POST", "/v1/leaderboards/" + key_ + "/scores", body,
                       [](const json::Value& j) {
